+++ packages/proxy/jest.config.js
@@ -4,9 +4,9 @@ module.exports = Object.assign({}, config, {
   coverageThreshold: {
     global: {
       branches: 79,
-      functions: 94,
-      lines: 87,
-      statements: 87,
+      functions: 90,
+      lines: 86,
+      statements: 86,
     },
   },
 });
