+++ packages/proxy/src/index.ts
@@ -1,1 +1,2 @@
 export * from './proxy';
+export * from './uplink-util';
