+++ packages/search-indexer/.babelrc
@@ -1,0 +1,3 @@
+{
+  "extends": "../../.babelrc"
+}
