+++ packages/search-indexer/jest.config.js
@@ -3,9 +3,10 @@ const config = require('../../jest/config');
 module.exports = Object.assign({}, config, {
   coverageThreshold: {
     global: {
-      branches: 0,
-      functions: 0,
-      lines: 0,
+      branches: 79,
+      functions: 94,
+      lines: 87,
+      statements: 87,
     },
   },
 });
