+++ packages/search-indexer/src/index.ts
@@ -1,0 +1,1 @@
+export { default as SearchMemoryIndexer } from './indexer';
