+++ packages/search-indexer/src/indexer.ts
