+++ packages/search-indexer/tsconfig.json
@@ -11,6 +11,12 @@
       "path": "../config"
     },
     {
+      "path": "../core/core"
+    },
+    {
+      "path": "../logger/logger"
+    },
+    {
       "path": "../utils"
     }
   ]
