+++ packages/search/src/index.ts
@@ -1,1 +1,3 @@
-export { default as SearchMemoryIndexer } from './indexer';
+export { Search as default } from './search';
+export * from './search-utils';
+export * from './transform-results';
