+++ packages/store/src/index.ts
@@ -1,6 +1,5 @@
 export { Storage } from './storage';
 export * from './lib/storage-utils';
-export * from './lib/search-utils';
 export * from './lib/versions-utils';
 export * from './lib/star-utils';
 export * from './type';
